Modify the Employee

+++ src/main/java/danang03/STBackend/domain/employee/EmployeeService.java
@@ -1,9 +1,16 @@
 package danang03.STBackend.domain.employee;
 
 import danang03.STBackend.domain.employee.dto.AddEmployeeRequest;
+import danang03.STBackend.domain.employee.dto.EmployeeDetailResponse;
 import danang03.STBackend.domain.employee.dto.EmployeeResponse;
 import danang03.STBackend.domain.employee.dto.UpdateEmployeeRequest;
 import danang03.STBackend.domain.image.S3Service;
+import danang03.STBackend.domain.projects.EmployeeProject;
+import danang03.STBackend.domain.projects.EmployeeProjectRepository;
+import danang03.STBackend.domain.projects.Project;
+import danang03.STBackend.domain.projects.dto.ProjectResponse;
+import java.util.List;
+import org.aspectj.weaver.ast.Literal;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -16,12 +23,14 @@ import org.springframework.web.multipart.MultipartFile;
 @Service
 public class EmployeeService {
     private final EmployeeRepository employeeRepository;
+    private final EmployeeProjectRepository employeeProjectRepository;
     private final S3Service s3Service;
 
 
     @Autowired
-    public EmployeeService(EmployeeRepository employeeRepository, S3Service s3Service) {
+    public EmployeeService(EmployeeRepository employeeRepository, EmployeeProjectRepository employeeProjectRepository, S3Service s3Service) {
         this.employeeRepository = employeeRepository;
+        this.employeeProjectRepository = employeeProjectRepository;
         this.s3Service = s3Service;
     }
 
@@ -58,6 +67,25 @@ public class EmployeeService {
                 .joiningDate(employee.getJoiningDate())
                 .role(employee.getRole())
                 .imageUrl(employee.getImageUrl()).build();
+    }
+
+    public EmployeeDetailResponse getEmployeeDetail(Long employeeId) {
+        EmployeeResponse employeeResponse = getEmployee(employeeId);
+        List<EmployeeProject> employeeProjects = employeeProjectRepository.findByEmployeeId(employeeId);
+        List<ProjectResponse> projectResponses = employeeProjects.stream()
+                .map(employeeProject -> {
+                    Project project = employeeProject.getProject();
+                        return new ProjectResponse(
+                                project.getId(),
+                                project.getName(),
+                                project.getDescription(),
+                                project.getStartDate(),
+                                project.getEndDate(),
+                                project.getStatus()
+                        );
+                })
+                .toList();
+        return new EmployeeDetailResponse(employeeResponse, projectResponses);
     }
 
     public Page<EmployeeResponse> getEmployeesByPage(Pageable pageable) {
@@ -101,6 +129,12 @@ public class EmployeeService {
     public void deleteEmployee(Long id) {
         if (!employeeRepository.existsById(id)) {
             throw new IllegalArgumentException("employee with id " + id + " not found");
+        }
+
+        // check employee in project
+        boolean hasProjects = employeeProjectRepository.existsByEmployeeId(id);
+        if (hasProjects) {
+            throw new IllegalStateException("Employee with id " + id + " cannot be deleted because it has associated projects.");
         }
         employeeRepository.deleteById(id);
     }
